# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bureau Agent API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /calls:
    get:
      operationId: listCalls
      tags: [calls]
      summary: List all calls
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [answered, missed, voicemail, outgoing, all]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of calls
          content:
            application/json:
              schema:
                type: object
                properties:
                  calls:
                    type: array
                    items:
                      $ref: "#/components/schemas/Call"
                  total:
                    type: integer
                required: [calls, total]
    post:
      operationId: createCall
      tags: [calls]
      summary: Create a new call record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCallBody"
      responses:
        "201":
          description: Call created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
  /calls/{id}:
    get:
      operationId: getCall
      tags: [calls]
      summary: Get a single call
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Call details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
        "404":
          description: Not found
    patch:
      operationId: updateCall
      tags: [calls]
      summary: Update a call record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCallBody"
      responses:
        "200":
          description: Updated call
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Call"
    delete:
      operationId: deleteCall
      tags: [calls]
      summary: Delete a call record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # === CONTACTS ===
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List contacts
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
            enum: [client, prospect, fournisseur, partenaire, autre, all]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: object
                properties:
                  contacts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Contact"
                  total:
                    type: integer
                required: [contacts, total]
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactBody"
      responses:
        "201":
          description: Contact created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    get:
      operationId: getContact
      tags: [contacts]
      summary: Get a single contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contact details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactBody"
      responses:
        "200":
          description: Updated contact
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # === TASKS ===
...
components:
  schemas:
...
    Call:
      type: object
      properties:
        id:
          type: integer
        contactId:
          type: integer
          nullable: true
        contactName:
          type: string
          nullable: true
        phoneNumber:
          type: string
        direction:
          type: string
          enum: [entrant, sortant]
        status:
          type: string
          enum: [repondu, manque, messagerie, en_cours]
        duration:
          type: integer
          description: Duration in seconds
        notes:
          type: string
          nullable: true
        sentiment:
          type: string
          enum: [positif, neutre, negatif]
          nullable: true
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, phoneNumber, direction, status, duration, tags, createdAt, updatedAt]
    CreateCallBody:
      type: object
      properties:
        contactId:
          type: integer
          nullable: true
        phoneNumber:
          type: string
        direction:
          type: string
          enum: [entrant, sortant]
        status:
          type: string
          enum: [repondu, manque, messagerie, en_cours]
        duration:
          type: integer
        notes:
          type: string
          nullable: true
        sentiment:
          type: string
          enum: [positif, neutre, negatif]
          nullable: true
        tags:
          type: array
          items:
            type: string
      required: [phoneNumber, direction, status, duration]
    UpdateCallBody:
      type: object
      properties:
        contactId:
          type: integer
          nullable: true
        status:
          type: string
          enum: [repondu, manque, messagerie, en_cours]
        duration:
          type: integer
        notes:
          type: string
          nullable: true
        sentiment:
          type: string
          enum: [positif, neutre, negatif]
          nullable: true
        tags:
          type: array
          items:
            type: string
    Contact:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
        mobile:
          type: string
          nullable: true
        category:
          type: string
          enum: [client, prospect, fournisseur, partenaire, autre]
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        totalCalls:
          type: integer
        lastCallAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, firstName, lastName, phone, category, totalCalls, createdAt, updatedAt]
    CreateContactBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
        mobile:
          type: string
          nullable: true
        category:
          type: string
          enum: [client, prospect, fournisseur, partenaire, autre]
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [firstName, lastName, phone, category]
    UpdateContactBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        company:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
        mobile:
          type: string
          nullable: true
        category:
          type: string
          enum: [client, prospect, fournisseur, partenaire, autre]
        address:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...