# lib/api-zod/src/generated/types/call.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bureau Agent API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CallDirection } from "./callDirection";
import type { CallSentiment } from "./callSentiment";
import type { CallStatus } from "./callStatus";

export interface Call {
  id: number;
  contactId?: number | null;
  contactName?: string | null;
  phoneNumber: string;
  direction: CallDirection;
  status: CallStatus;
  /** Duration in seconds */
  duration: number;
  notes?: string | null;
  sentiment?: CallSentiment;
  tags: string[];
  createdAt: Date;
  updatedAt: Date;
}