# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bureau Agent API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ContactCategory } from "./contactCategory";

export interface Contact {
  id: number;
  firstName: string;
  lastName: string;
  company?: string | null;
  email?: string | null;
  phone: string;
  mobile?: string | null;
  category: ContactCategory;
  address?: string | null;
  notes?: string | null;
  totalCalls: number;
  lastCallAt?: Date | null;
  createdAt: Date;
  updatedAt: Date;
}