# lib/api-zod/src/generated/types/createContactBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bureau Agent API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContactBodyCategory } from "./createContactBodyCategory";

export interface CreateContactBody {
  firstName: string;
  lastName: string;
  company?: string | null;
  email?: string | null;
  phone: string;
  mobile?: string | null;
  category: CreateContactBodyCategory;
  address?: string | null;
  notes?: string | null;
}